+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support JSON as input
+- Preserve whitespace with optional wrapper
 
 ### Changed
 - Rename to `DeltaHtml.to_html`